pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-organizer"
version = "0.1.0"
description = "Tool to organize photos by detecting duplicates and organizing by date"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
]

[project.scripts]
image-organizer = "main:main"
